rename CrossJoin to NLJoin
