+++ be/src/exec/pipeline/nljoin/nljoin_build_operator.h
@@ -5,7 +5,7 @@
 #include <utility>
 
 #include "column/vectorized_fwd.h"
-#include "exec/pipeline/crossjoin/cross_join_context.h"
+#include "exec/pipeline/nljoin/nljoin_context.h"
 #include "exec/pipeline/operator.h"
 
 namespace starrocks::pipeline {
@@ -15,7 +15,7 @@ namespace starrocks::pipeline {
 class NLJoinBuildOperator final : public Operator {
 public:
     NLJoinBuildOperator(OperatorFactory* factory, int32_t id, int32_t plan_node_id, const int32_t driver_sequence,
-                        const std::shared_ptr<CrossJoinContext>& cross_join_context)
+                        const std::shared_ptr<NLJoinContext>& cross_join_context)
             : Operator(factory, id, "nestloop_join_build", plan_node_id, driver_sequence),
               _cross_join_context(cross_join_context) {
         _cross_join_context->ref();
@@ -40,7 +40,22 @@ public:
 private:
     bool _is_finished = false;
 
-    const std::shared_ptr<CrossJoinContext>& _cross_join_context;
+    const std::shared_ptr<NLJoinContext>& _cross_join_context;
+};
+
+class NLJoinBuildOperatorFactory final : public OperatorFactory {
+public:
+    NLJoinBuildOperatorFactory(int32_t id, int32_t plan_node_id, std::shared_ptr<NLJoinContext> cross_join_context)
+            : OperatorFactory(id, "nljoin_build", plan_node_id), _cross_join_context(std::move(cross_join_context)) {}
+
+    ~NLJoinBuildOperatorFactory() override = default;
+
+    OperatorPtr create(int32_t degree_of_parallelism, int32_t driver_sequence) override {
+        return std::make_shared<NLJoinBuildOperator>(this, _id, _plan_node_id, driver_sequence, _cross_join_context);
+    }
+
+private:
+    std::shared_ptr<NLJoinContext> _cross_join_context;
 };
 
 } // namespace starrocks::pipeline
