+++ be/src/exec/vectorized/hash_join_node.h
@@ -73,8 +73,6 @@ private:
     Status _do_publish_runtime_filters(RuntimeState* state, int64_t limit);
     Status _push_down_in_filter(RuntimeState* state);
 
-    static std::string _get_join_type_str(TJoinOp::type join_type);
-
     friend ExecNode;
     // _hash_join_node is used to construct HashJoiner, the reference is sound since
     // it's only used in FragmentExecutor::prepare function.
